cr feedback, whitespace, newlines at file ends, warning fixed
