+++ lib/shapely/impl/src/generator.rs
@@ -9,6 +9,8 @@ use std::ops::Generator;
 use std::ops::GeneratorState;
 use std::pin::Pin;
 
+
+
 // ==========================
 // === GeneratingIterator ===
 // ==========================
