+++ lib/shapely/macros/src/derive_iterator.rs
@@ -8,6 +8,8 @@ use macro_utils::ty_path_type_args;
 use macro_utils::variant_depends_on;
 use inflector::Inflector;
 
+
+
 // =============
 // === IsMut ===
 // =============
